Modify refactor move SPU classes in the right package
