+++ engine/src/main/java/it/unibo/arces/wot/sepa/engine/processing/subscriptions/ISubscriptionProcUnit.java
@@ -7,7 +7,7 @@ import it.unibo.arces.wot.sepa.commons.sparql.BindingsResults;
 public interface ISubscriptionProcUnit extends Runnable {
 
     Response init();
-
+    //TODO: refactor getCurrentResults
     BindingsResults getFirstResults();
 
     void terminate();
