+++ engine/src/main/java/it/unibo/arces/wot/sepa/engine/processing/subscriptions/SPUEndOfProcessing.java
@@ -1,4 +1,4 @@
-package it.unibo.arces.wot.sepa.engine.processing;
+package it.unibo.arces.wot.sepa.engine.processing.subscriptions;
 
 import it.unibo.arces.wot.sepa.commons.response.Response;
