+++ engine/src/main/java/it/unibo/arces/wot/sepa/engine/processing/subscriptions/SPUProcessingResultsQueue.java
@@ -1,4 +1,6 @@
-package it.unibo.arces.wot.sepa.engine.processing;
+package it.unibo.arces.wot.sepa.engine.processing.subscriptions;
+
+import it.unibo.arces.wot.sepa.engine.processing.subscriptions.SPU;
 
 import java.util.concurrent.ConcurrentLinkedQueue;
